+++ src/lib.rs
@@ -116,18 +116,18 @@ impl Clock {
         enable();
 
         // Initialize the RTC itself.
-        reset();
+        reset()?;
         // If the power bit is active, we need to reset.
         let status = try_read_status()?;
         if status.contains(&Status::POWER) {
-            reset();
+            reset()?;
         }
         // If we are in test mode, we need to reset.
-        if is_test_mode() {
-            reset();
+        if is_test_mode()? {
+            reset()?;
         }
         // Set to 24-hour time.
-        set_status(Status::HOUR_24);
+        set_status(Status::HOUR_24)?;
 
         let rtc_offset = try_read_datetime_offset()?;
 
@@ -421,7 +421,12 @@ impl<'de> Deserialize<'de> for Clock {
         if result.is_ok() {
             // Enable operations with the RTC via General Purpose I/O (GPIO).
             enable();
-            set_status(Status::HOUR_24);
+            set_status(Status::HOUR_24).map_err(|error| {
+                de::Error::custom(format_args!(
+                    "could not set RTC status 24 hour bit: {}",
+                    error
+                ))
+            })?;
             // If the power bit is active, the clock is unreadable.
             let status = try_read_status().map_err(|error| {
                 de::Error::custom(format_args!("could not read RTC status: {}", error))
@@ -432,7 +437,12 @@ impl<'de> Deserialize<'de> for Clock {
                 ));
             }
             // If we are in test mode, the clock is unreadable.
-            if is_test_mode() {
+            if is_test_mode().map_err(|error| {
+                de::Error::custom(format_args!(
+                    "could not detect if RTC is in test mode: {}",
+                    error
+                ))
+            })? {
                 return Err(de::Error::custom("RTC is in test mode and must be reset"));
             }
         }
@@ -450,14 +460,17 @@ pub fn main() {
 #[cfg(test)]
 mod tests {
     use super::{
+        gpio,
         Clock,
         Error,
     };
+    use crate::date_time::RtcDateTimeOffset;
     use claims::{
         assert_err_eq,
         assert_ok,
         assert_ok_eq,
     };
+    use deranged::RangedU32;
     use gba_test::test;
     use time_macros::{
         date,
@@ -470,11 +483,8 @@ mod tests {
         not(no_rtc),
         ignore = "This test requires the RTC to be disabled. Ensure no RTC is configured and pass `--cfg no_rtc` to enable."
     )]
-    fn new_clock_invalid_offset() {
-        assert_err_eq!(
-            Clock::new(datetime!(2012-12-21 5:23)),
-            Error::InvalidMonth(0)
-        );
+    fn new_clock_not_enabled() {
+        assert_err_eq!(Clock::new(datetime!(2012-12-21 5:23)), Error::NotEnabled);
     }
 
     #[test]
@@ -487,6 +497,38 @@ mod tests {
         let clock = assert_ok!(Clock::new(datetime));
 
         assert_ok_eq!(clock.read_datetime(), datetime);
+    }
+
+    #[test]
+    #[cfg_attr(
+        not(no_rtc),
+        ignore = "This test requires the RTC to be disabled. Ensure no RTC is configured and pass `--cfg no_rtc` to enable."
+    )]
+    fn read_datetime_not_enabled() {
+        // Manually enable RTC.
+        gpio::enable();
+        // Manually construct a `Clock` object, despite RTC being disabled.
+        //
+        // This is to simulate an RTC failing after `Clock` construction.
+        let clock = Clock {
+            base_date: date!(2012 - 12 - 21),
+            rtc_offset: RtcDateTimeOffset(RangedU32::new_static::<0>()),
+        };
+
+        assert_err_eq!(clock.read_datetime(), Error::NotEnabled);
+    }
+
+    #[test]
+    #[cfg_attr(
+        not(rtc),
+        ignore = "This test requires a functioning RTC. Ensure an RTC is configured and pass `--cfg rtc` to enable."
+    )]
+    fn read_datetime_after_disabled() {
+        let clock = assert_ok!(Clock::new(datetime!(2012-12-21 5:23)));
+
+        gpio::disable();
+
+        assert_err_eq!(clock.read_datetime(), Error::NotEnabled);
     }
 
     #[test]
@@ -505,6 +547,42 @@ mod tests {
 
     #[test]
     #[cfg_attr(
+        not(no_rtc),
+        ignore = "This test requires the RTC to be disabled. Ensure no RTC is configured and pass `--cfg no_rtc` to enable."
+    )]
+    fn write_datetime_not_enabled() {
+        // Manually enable RTC.
+        gpio::enable();
+        // Manually construct a `Clock` object, despite RTC being disabled.
+        //
+        // This is to simulate an RTC failing after `Clock` construction.
+        let mut clock = Clock {
+            base_date: date!(2012 - 12 - 21),
+            rtc_offset: RtcDateTimeOffset(RangedU32::new_static::<0>()),
+        };
+
+        assert_err_eq!(
+            clock.write_datetime(datetime!(2012-12-21 5:23)),
+            Error::NotEnabled
+        );
+    }
+
+    #[test]
+    #[cfg_attr(
+        not(rtc),
+        ignore = "This test requires a functioning RTC. Ensure an RTC is configured and pass `--cfg rtc` to enable."
+    )]
+    fn write_datetime_after_disabled() {
+        let datetime = datetime!(2012-12-21 5:23);
+        let mut clock = assert_ok!(Clock::new(datetime));
+
+        gpio::disable();
+
+        assert_err_eq!(clock.write_datetime(datetime), Error::NotEnabled);
+    }
+
+    #[test]
+    #[cfg_attr(
         not(rtc),
         ignore = "This test requires a functioning RTC. Ensure an RTC is configured and pass `--cfg rtc` to enable."
     )]
@@ -513,6 +591,38 @@ mod tests {
         let clock = assert_ok!(Clock::new(datetime));
 
         assert_ok_eq!(clock.read_date(), datetime.date());
+    }
+
+    #[test]
+    #[cfg_attr(
+        not(no_rtc),
+        ignore = "This test requires the RTC to be disabled. Ensure no RTC is configured and pass `--cfg no_rtc` to enable."
+    )]
+    fn read_date_not_enabled() {
+        // Manually enable RTC.
+        gpio::enable();
+        // Manually construct a `Clock` object, despite RTC being disabled.
+        //
+        // This is to simulate an RTC failing after `Clock` construction.
+        let clock = Clock {
+            base_date: date!(2012 - 12 - 21),
+            rtc_offset: RtcDateTimeOffset(RangedU32::new_static::<0>()),
+        };
+
+        assert_err_eq!(clock.read_date(), Error::NotEnabled);
+    }
+
+    #[test]
+    #[cfg_attr(
+        not(rtc),
+        ignore = "This test requires a functioning RTC. Ensure an RTC is configured and pass `--cfg rtc` to enable."
+    )]
+    fn read_date_after_disabled() {
+        let clock = assert_ok!(Clock::new(datetime!(2012-12-21 5:23)));
+
+        gpio::disable();
+
+        assert_err_eq!(clock.read_date(), Error::NotEnabled);
     }
 
     #[test]
@@ -530,6 +640,38 @@ mod tests {
 
     #[test]
     #[cfg_attr(
+        not(no_rtc),
+        ignore = "This test requires the RTC to be disabled. Ensure no RTC is configured and pass `--cfg no_rtc` to enable."
+    )]
+    fn write_date_not_enabled() {
+        // Manually enable RTC.
+        gpio::enable();
+        // Manually construct a `Clock` object, despite RTC being disabled.
+        //
+        // This is to simulate an RTC failing after `Clock` construction.
+        let mut clock = Clock {
+            base_date: date!(2012 - 12 - 21),
+            rtc_offset: RtcDateTimeOffset(RangedU32::new_static::<0>()),
+        };
+
+        assert_err_eq!(clock.write_date(date!(2012 - 12 - 21)), Error::NotEnabled);
+    }
+
+    #[test]
+    #[cfg_attr(
+        not(rtc),
+        ignore = "This test requires a functioning RTC. Ensure an RTC is configured and pass `--cfg rtc` to enable."
+    )]
+    fn write_date_after_disabled() {
+        let mut clock = assert_ok!(Clock::new(datetime!(2000-01-01 0:00)));
+
+        gpio::disable();
+
+        assert_err_eq!(clock.write_date(date!(2012 - 12 - 21)), Error::NotEnabled);
+    }
+
+    #[test]
+    #[cfg_attr(
         not(rtc),
         ignore = "This test requires a functioning RTC. Ensure an RTC is configured and pass `--cfg rtc` to enable."
     )]
@@ -538,6 +680,38 @@ mod tests {
         let clock = assert_ok!(Clock::new(datetime));
 
         assert_ok_eq!(clock.read_time(), datetime.time());
+    }
+
+    #[test]
+    #[cfg_attr(
+        not(no_rtc),
+        ignore = "This test requires the RTC to be disabled. Ensure no RTC is configured and pass `--cfg no_rtc` to enable."
+    )]
+    fn read_time_not_enabled() {
+        // Manually enable RTC.
+        gpio::enable();
+        // Manually construct a `Clock` object, despite RTC being disabled.
+        //
+        // This is to simulate an RTC failing after `Clock` construction.
+        let clock = Clock {
+            base_date: date!(2012 - 12 - 21),
+            rtc_offset: RtcDateTimeOffset(RangedU32::new_static::<0>()),
+        };
+
+        assert_err_eq!(clock.read_time(), Error::NotEnabled);
+    }
+
+    #[test]
+    #[cfg_attr(
+        not(rtc),
+        ignore = "This test requires a functioning RTC. Ensure an RTC is configured and pass `--cfg rtc` to enable."
+    )]
+    fn read_time_after_disabled() {
+        let clock = assert_ok!(Clock::new(datetime!(2012-12-21 5:23)));
+
+        gpio::disable();
+
+        assert_err_eq!(clock.read_time(), Error::NotEnabled);
     }
 
     #[test]
@@ -551,5 +725,37 @@ mod tests {
         assert_ok!(clock.write_time(time!(22:22)));
 
         assert_ok_eq!(clock.read_datetime(), datetime!(2012-12-21 22:22));
+    }
+
+    #[test]
+    #[cfg_attr(
+        not(no_rtc),
+        ignore = "This test requires the RTC to be disabled. Ensure no RTC is configured and pass `--cfg no_rtc` to enable."
+    )]
+    fn write_time_not_enabled() {
+        // Manually enable RTC.
+        gpio::enable();
+        // Manually construct a `Clock` object, despite RTC being disabled.
+        //
+        // This is to simulate an RTC failing after `Clock` construction.
+        let mut clock = Clock {
+            base_date: date!(2012 - 12 - 21),
+            rtc_offset: RtcDateTimeOffset(RangedU32::new_static::<0>()),
+        };
+
+        assert_err_eq!(clock.write_time(time!(22:22)), Error::NotEnabled);
+    }
+
+    #[test]
+    #[cfg_attr(
+        not(rtc),
+        ignore = "This test requires a functioning RTC. Ensure an RTC is configured and pass `--cfg rtc` to enable."
+    )]
+    fn write_time_after_disabled() {
+        let mut clock = assert_ok!(Clock::new(datetime!(2000-01-01 0:00)));
+
+        gpio::disable();
+
+        assert_err_eq!(clock.write_time(time!(22:22)), Error::NotEnabled);
     }
 }